refactor: move octokit loader into shared helper
